pyproject.toml
[tool.poetry]
name = "grants-tagger"
version = "0.1.0"
description = "A machine learning model to tag grants. You can find more information on https://github.com/wellcometrust/grants_tagger"
authors = ["Arne Robben <a.robben@wellcome.ac.uk>"]
readme = "README.md"
packages = [{include = "grants_tagger"}]

[tool.poetry.dependencies]
python = "^3.8"
xlrd = "^2.0.1"
numpy = "1.21"
pandas = "1.3.5"
typer = "^0.9.0"
PyYAML = "^6.0"
tqdm = "^4.64.0"
requests = "^2.28.1"
scikit-learn = "1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
libpecos = "0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
grants_tagger = "grants_tagger.__main__:app"